s2-codeservice-template/pyproject.toml
[build-system]
  build-backend = "poetry.core.masonry.api"
  requires = ["poetry-core"]

[project]
  dependencies = ["singlestoredb (>=1.11.0,<2.0.0)","fastapi (>=0.115.8,<0.116.0)","python-dotenv (>=1.0.1,<2.0.0)", "pydantic (>=2.10.6,<3.0.0)", "uvicorn (>=0.34.0,<0.35.0)"]
  description = ""
  name = "s2-codeservice-template"
  readme = "README.md"
  requires-python = "^3.11" 
  version = "1.0.0"
  [project.scripts]
    nova = "s2_codeservice_template:main"
    local-dev = "s2_codeservice_template:run_on_local"
